+++ cmta/src/test/scala/cmt/admin/cli/CommandLineParseSpec.scala
@@ -1,9 +1,10 @@
-package cmt
+package cmt.admin.cli
 
 import cmt.admin.cli.{AdminCliParser, CliOptions}
 import cmt.support.CommandLineParseTestBase
+import cmt.*
 
-class CommandLineParseTest
+class CommandLineParseSpec
     extends CommandLineParseTestBase[CliOptions](
       AdminCliParser.parse,
       StudentifyArguments,
