+++ cmta/src/test/scala/cmt/admin/cli/DelinearizeArguments.scala
@@ -1,16 +1,11 @@
-package cmt
+package cmt.admin.cli
 
 import cmt.TestDirectories
 import cmt.admin.Domain.{LinearizeBaseDirectory, MainRepository}
 import cmt.admin.cli.CliCommand.DeLinearize
-import cmt.admin.cli.CliOptions
 import cmt.support.CommandLineArguments
-import cmt.support.CommandLineArguments.*
-import org.scalatest.BeforeAndAfterAll
-import org.scalatest.matchers.should.Matchers
+import cmt.support.CommandLineArguments.{invalidArgumentsTable, validArgumentsTable}
 import org.scalatest.prop.Tables
-import org.scalatest.wordspec.AnyWordSpecLike
-import sbt.io.IO
 import sbt.io.syntax.{File, file}
 import scopt.OEffect.ReportError
